fix: missing msgctxt in parser
The issue was caused by a mistake of using the children array
index to access the translation function arguments. This commit
fixes the issue by using an individual variable to store the index
of the translation key.

Also using the node type to make sure it is a string, and unwrap
the PHP argument node.

Fixes #11